backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
distro==1.9.0
elevenlabs==2.3.0
exceptiongroup==1.3.0
fastapi==0.115.12
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
openai==1.86.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
